[#4] Added `Facilities` which contain doctors (`Users` with specific `Role`).

+++ backend/src/roles/roles.service.ts
@@ -1,13 +1,17 @@
-import { Injectable, NotFoundException } from '@nestjs/common';
+import { BadRequestException, Injectable, NotFoundException } from '@nestjs/common';
 import { User } from 'src/users/entities/user.entity';
+import { UsersService } from 'src/users/users.service';
 import { AssociateRoleDto } from './dto/associate-role.dto';
 import { CreateRoleDto } from './dto/create-role.dto';
 import { DissolveRoleDto } from './dto/dissolve-role.dto';
-import { ResponseRoleDto } from './dto/response-role.dto';
-import { Role, UserRole } from './entities/role.entity';
+import { RoleType } from './entities/role-types.entity';
+import { UserRole } from './entities/role-user.entity';
+import { Role } from './entities/role.entity';
 
 @Injectable()
 export class RolesService {
+  constructor(private usersService: UsersService) { }
+
   create(createRoleDto: CreateRoleDto) {
     return Role.create(createRoleDto);
   }
@@ -17,7 +21,7 @@ export class RolesService {
     return Role.findAll({ include: includeOptions });
   }
 
-  findOne(uuid: string, includeUsers: boolean = false) {
+  findOne(uuid: string, includeUsers: boolean = false): Promise<Role | undefined> {
     const includeOptions = includeUsers ? [User] : [];
     return Role.findOne({
       where: { uuid: uuid },
@@ -42,7 +46,7 @@ export class RolesService {
   async associate(associateRole: AssociateRoleDto): Promise<UserRole> {
     return Promise.all(
       [
-        User.findOne({ where: { uuid: associateRole.userUuid } }),
+        this.usersService.findOne(associateRole.userUuid),
         Role.findOne({ where: { uuid: associateRole.roleUuid } })
       ]).then(([user, role]) => {
         if (!user || !role) {
@@ -55,11 +59,15 @@ export class RolesService {
   async dissolve(dissolveRole: DissolveRoleDto): Promise<number> {
     return Promise.all(
       [
-        User.findOne({ where: { uuid: dissolveRole.userUuid } }),
+        this.usersService.findOne(dissolveRole.userUuid),
         Role.findOne({ where: { uuid: dissolveRole.roleUuid } })
       ]).then(([user, role]) => {
         if (!user || !role) {
           throw new NotFoundException('Invalid user or role!');
+        }
+        // Forbid removing role 'doctor' from user who is assigned to any facility.
+        if (role.name === RoleType.Doctor && user.facilities.length > 0) {
+          throw new BadRequestException(`User is a ${role.name} assigned to facilities, role "${role.name}" can not be removed!`);
         }
         return UserRole.destroy({ where: { userId: user.id, roleId: role.id } });
       })
